[REFACTOR] Split game logic actions into individual files
