+++ pandora-common/src/gameLogic/actionLogic/index.ts
@@ -1,3 +1,4 @@
+export * from './actions/_index';
 export * from './appearanceActionProblems';
 export * from './appearanceActionProcessingContext';
 export * from './appearanceActions';
